ci: port move operate importer tests to nightly job
