+++ operate/pom.xml
@@ -58,7 +58,7 @@
     <license.header>${project.baseUri}/../../license/license-header</license.header>
 
     <includeITNames>**/*IT.java</includeITNames>
-    <excludeITNames>**/*5IT.java</excludeITNames>
+    <excludeITNames>**/OpenSearch*IT.java, **/Opensearch*IT.java</excludeITNames>
     <skipSurefire>false</skipSurefire>
     <failsafeForkCount>${env.LIMITS_CPU}</failsafeForkCount>
 
@@ -977,10 +977,10 @@
 
   <profiles>
     <profile>
-      <id>itStage1</id>
+      <id>operateItStage1</id>
       <properties>
         <includeITNames>**/*IT.java</includeITNames>
-        <excludeITNames>**/*ZeebeIT.java</excludeITNames>
+        <excludeITNames>**/*ZeebeIT.java, **/*ZeebeImportIT.java, **/OpenSearch*IT.java, **/Opensearch*IT.java</excludeITNames>
         <failsafeForkCount>1</failsafeForkCount>
       </properties>
     </profile>
@@ -990,7 +990,17 @@
       <properties>
         <includeITNames>**/*ZeebeIT.java</includeITNames>
         <excludeITNames>none</excludeITNames>
-        <failsafeForkCount>${env.LIMITS_CPU}</failsafeForkCount>
+        <failsafeForkCount>1</failsafeForkCount>
+        <skipSurefire>true</skipSurefire>
+      </properties>
+    </profile>
+
+    <profile>
+      <id>operateItImport</id>
+      <properties>
+        <includeITNames>**/*ZeebeImportIT.java</includeITNames>
+        <excludeITNames>none</excludeITNames>
+        <failsafeForkCount>1</failsafeForkCount>
         <skipSurefire>true</skipSurefire>
       </properties>
     </profile>
