+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/zeebeimport/FlowNodeMetadataZeebeImportIT.java
@@ -58,7 +58,7 @@ import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.servlet.MvcResult;
 
-public class FlowNodeMetadataZeebeIT extends OperateZeebeAbstractIT {
+public class FlowNodeMetadataZeebeImportIT extends OperateZeebeAbstractIT {
 
   @Autowired private CancelProcessInstanceHandler cancelProcessInstanceHandler;
