+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/zeebeimport/IdempotencyZeebeImportIT.java
@@ -16,62 +16,39 @@
  */
 package io.camunda.operate.zeebeimport;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import io.camunda.operate.entities.ProcessEntity;
-import io.camunda.operate.schema.indices.ProcessIndex;
-import io.camunda.operate.util.j5templates.OperateZeebeSearchAbstractIT;
-import java.io.IOException;
-import java.util.List;
-import org.junit.jupiter.api.Test;
+import io.camunda.operate.property.OperateProperties;
+import io.camunda.operate.util.TestApplication;
+import io.camunda.operate.util.apps.idempotency.ZeebeImportIdempotencyTestConfig;
+import io.camunda.operate.zeebe.ImportValueType;
+import java.util.function.Predicate;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 
-public class ProcessZeebeRefactoredIT extends OperateZeebeSearchAbstractIT {
+/**
+ * Tests that even if the Zeebe data is imported twice, in Operate Elasticsearch is is still
+ * consistent.
+ */
+@SpringBootTest(
+    classes = {ZeebeImportIdempotencyTestConfig.class, TestApplication.class},
+    properties = {
+      OperateProperties.PREFIX + ".importer.startLoadingDataOnStartup = false",
+      OperateProperties.PREFIX + ".archiver.rolloverEnabled = false",
+      "spring.main.allow-bean-definition-overriding=true",
+      "spring.mvc.pathmatch.matching-strategy=ANT_PATH_MATCHER"
+    })
+public class IdempotencyZeebeImportIT extends BasicZeebeImportIT {
 
-  @Autowired private ProcessIndex processIndex;
+  @Autowired
+  private ZeebeImportIdempotencyTestConfig.CustomElasticsearchBulkProcessor
+      elasticsearchBulkProcessor;
 
-  @Test
-  public void shouldReadFromMultiProcessDiagram() throws IOException {
-
-    // given
-    final String bpmnFile = "multi-process.bpmn";
-    operateTester.deployProcessAndWait(bpmnFile);
-
-    // when
-    final List<ProcessEntity> processEntities =
-        testSearchRepository.searchAll(processIndex.getAlias(), ProcessEntity.class);
-    final ProcessEntity process1 =
-        processEntities.stream()
-            .filter(x -> "process1".equals(x.getBpmnProcessId()))
-            .findAny()
-            .orElse(null);
-    final ProcessEntity process2 =
-        processEntities.stream()
-            .filter(x -> "process2".equals(x.getBpmnProcessId()))
-            .findAny()
-            .orElse(null);
-
-    // then
-    assertThat(processEntities).hasSize(2);
-
-    assertThat(process1).isNotNull();
-    assertThat(process1.getName()).isEqualTo("Process 1");
-    assertThat(process1.getResourceName()).isEqualTo(bpmnFile);
-    assertThat(process1.getFlowNodes())
-        .extracting("id")
-        .containsExactlyInAnyOrder("startEvent1", "task1", "endEvent1");
-    assertThat(process1.getFlowNodes())
-        .extracting("name")
-        .containsExactlyInAnyOrder("Start Event 1", "Task 1", "End Event 1");
-
-    assertThat(process2).isNotNull();
-    assertThat(process2.getName()).isEqualTo("Process 2");
-    assertThat(process2.getResourceName()).isEqualTo(bpmnFile);
-    assertThat(process2.getFlowNodes())
-        .extracting("id")
-        .containsExactlyInAnyOrder("startEvent2", "task2", "endEvent2");
-    assertThat(process2.getFlowNodes())
-        .extracting("name")
-        .containsExactlyInAnyOrder("Start Event 2", "Task 2", "End Event 2");
+  @Override
+  protected void processImportTypeAndWait(
+      final ImportValueType importValueType,
+      final Predicate<Object[]> waitTill,
+      final Object... arguments) {
+    searchTestRule.processRecordsWithTypeAndWait(importValueType, waitTill, arguments);
+    searchTestRule.processRecordsWithTypeAndWait(importValueType, waitTill, arguments);
+    elasticsearchBulkProcessor.cancelAttempts();
   }
 }
