+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/zeebeimport/ImportMultitenancyDisabledZeebeImportIT.java
@@ -17,21 +17,20 @@
 package io.camunda.operate.zeebeimport;
 
 import static io.camunda.operate.qa.util.RestAPITestUtil.createGetAllProcessInstancesRequest;
-import static io.camunda.operate.util.ThreadUtil.sleepFor;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.doReturn;
 
+import io.camunda.operate.entities.FlowNodeInstanceEntity;
 import io.camunda.operate.property.OperateProperties;
-import io.camunda.operate.util.*;
-import io.camunda.operate.util.searchrepository.TestSearchRepository;
+import io.camunda.operate.util.OperateZeebeAbstractIT;
+import io.camunda.operate.util.TestApplication;
+import io.camunda.operate.util.ZeebeTestUtil;
 import io.camunda.operate.webapp.reader.ListViewReader;
+import io.camunda.operate.webapp.rest.dto.listview.ListViewProcessInstanceDto;
+import io.camunda.operate.webapp.rest.dto.listview.ListViewRequestDto;
 import io.camunda.operate.webapp.rest.dto.listview.ListViewResponseDto;
-import io.camunda.zeebe.protocol.Protocol;
-import java.io.IOException;
-import java.util.ArrayList;
+import io.camunda.operate.webapp.security.tenant.TenantService;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.stream.Collectors;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -45,61 +44,50 @@ import org.springframework.test.context.junit4.SpringRunner;
       OperateProperties.PREFIX + ".importer.startLoadingDataOnStartup = false",
       OperateProperties.PREFIX + ".archiver.rolloverEnabled = false",
       "spring.mvc.pathmatch.matching-strategy=ANT_PATH_MATCHER",
-      // make batch size smaller
-      OperateProperties.PREFIX + ".zeebeElasticsearch.batchSize = 2"
+      OperateProperties.PREFIX + ".multiTenancy.enabled = false"
     })
-public class ImportWithMissedDataZeebeIT extends OperateZeebeAbstractIT {
-
-  @Autowired private TestSearchRepository searchRepository;
+public class ImportMultitenancyDisabledZeebeImportIT extends OperateZeebeAbstractIT {
 
   @Autowired private ListViewReader listViewReader;
 
+  private final String defaultTenantId = "<default>";
+
   @Test
-  public void testProcessInstanceCreated() throws IOException {
+  public void testDefaultTenantIsAssignedAndImported() {
+    doReturn(TenantService.AuthenticatedTenants.assignedTenants(List.of(defaultTenantId)))
+        .when(tenantService)
+        .getAuthenticatedTenants();
+
     // having
     final String processId = "demoProcess";
     final Long processDefinitionKey = deployProcess("demoProcess_v_1.bpmn");
-    final List<Long> processInstanceKeys = new ArrayList<>();
-    for (int i = 0; i < 20; i++) {
-      processInstanceKeys.add(
-          ZeebeTestUtil.startProcessInstance(zeebeClient, processId, "{\"a\": \"b\"}"));
-    }
-    // remove "middle part" of Zeebe data
-    // split by partitions
-    final Map<Integer, List<Long>> keysByPartition =
-        processInstanceKeys.stream()
-            .collect(Collectors.groupingBy(a -> Protocol.decodePartitionId(a)));
-    final List<Long> keysForDeletion = new ArrayList<>();
-    for (final Map.Entry<Integer, List<Long>> partition : keysByPartition.entrySet()) {
-      final AtomicInteger counter = new AtomicInteger();
-      final Map<Integer, List<Long>> splittedKeys =
-          partition.getValue().stream()
-              .collect(Collectors.groupingBy(x -> counter.getAndIncrement() / 2));
-      for (int j = 1; j < splittedKeys.size() - 1; j++) {
-        keysForDeletion.addAll(splittedKeys.get(j));
-      }
-    }
 
-    // wait for Zeebe to export data
-    sleepFor(3000L);
-    searchRepository.deleteByTermsQuery(
-        zeebeRule.getPrefix() + "*", "value.processInstanceKey", keysForDeletion);
-    searchTestRule.refreshZeebeIndices();
-    processInstanceKeys.removeAll(keysForDeletion);
+    // when
+    final Long processInstanceKey =
+        ZeebeTestUtil.startProcessInstance(zeebeClient, processId, "{\"a\": \"b\"}");
+    searchTestRule.processAllRecordsAndWait(flowNodeIsActiveCheck, processInstanceKey, "taskA");
 
-    // when importer runs
-    searchTestRule.processAllRecordsAndWait(processInstancesAreStartedCheck, processInstanceKeys);
+    // then
+    // assert process instance
+    final ListViewProcessInstanceDto pi = getSingleProcessInstanceForListView();
+    assertThat(pi.getTenantId()).isEqualTo(defaultTenantId);
+    // assert flow node instances
+    final List<FlowNodeInstanceEntity> allFlowNodeInstances =
+        tester.getAllFlowNodeInstances(processInstanceKey);
+    assertThat(allFlowNodeInstances).extracting("tenantId").containsOnly(defaultTenantId);
+    // assert variables
 
-    // then all expected instances are loaded - imported is not blocked
-    final ListViewResponseDto listViewResponseDto =
-        listViewReader.queryProcessInstances(
-            createGetAllProcessInstancesRequest(
-                q ->
-                    q.setIds(
-                        processInstanceKeys.stream()
-                            .map(String::valueOf)
-                            .collect(Collectors.toList()))));
+  }
 
-    assertThat(listViewResponseDto.getTotalCount()).isEqualTo(processInstanceKeys.size());
+  private ListViewProcessInstanceDto getSingleProcessInstanceForListView(
+      final ListViewRequestDto request) {
+    final ListViewResponseDto listViewResponse = listViewReader.queryProcessInstances(request);
+    assertThat(listViewResponse.getTotalCount()).isEqualTo(1);
+    assertThat(listViewResponse.getProcessInstances()).hasSize(1);
+    return listViewResponse.getProcessInstances().get(0);
+  }
+
+  private ListViewProcessInstanceDto getSingleProcessInstanceForListView() {
+    return getSingleProcessInstanceForListView(createGetAllProcessInstancesRequest());
   }
 }
