+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/zeebeimport/IntermediateThrowEventZeebeImportIT.java
@@ -26,23 +26,18 @@ import java.util.List;
 import org.junit.Test;
 import org.springframework.test.annotation.IfProfileValue;
 
-public class MessageEndEventZeebeIT extends OperateZeebeAbstractIT {
+public class IntermediateThrowEventZeebeImportIT extends OperateZeebeAbstractIT {
 
   @Test
   @IfProfileValue(name = "spring.profiles.active", value = "test")
-  public void shouldImportMessageEndEvent() {
-    final String flowNodeId = "messageEndEvent";
-    final String jobKey = "taskDefinition";
+  public void shouldImportIntermediateThrowEvent() {
     // given
     tester
-        .deployProcess("message-end-event.bpmn")
+        .deployProcess("intermediate-throw-event.bpmn")
         .waitUntil()
         .processIsDeployed()
         .then()
-        .startProcessInstance("message-end-event-process", null)
-        .flowNodeIsActive(flowNodeId)
-        .then()
-        .completeTask(flowNodeId, jobKey)
+        .startProcessInstance("intermediate-throw-event-process", null)
         .waitUntil()
         .processInstanceIsFinished();
 
@@ -51,6 +46,10 @@ public class MessageEndEventZeebeIT extends OperateZeebeAbstractIT {
         tester.getAllFlowNodeInstances(tester.getProcessInstanceKey());
     // then
     assertThat(map(flowNodes, FlowNodeInstanceEntity::getType))
-        .isEqualTo(List.of(FlowNodeType.START_EVENT, FlowNodeType.END_EVENT));
+        .isEqualTo(
+            List.of(
+                FlowNodeType.START_EVENT,
+                FlowNodeType.INTERMEDIATE_THROW_EVENT,
+                FlowNodeType.END_EVENT));
   }
 }
