+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/zeebeimport/MultiProcessZeebeImportIT.java
@@ -16,67 +16,62 @@
  */
 package io.camunda.operate.zeebeimport;
 
-import static io.camunda.operate.webapp.rest.ProcessInstanceRestService.PROCESS_INSTANCE_URL;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import io.camunda.operate.schema.templates.SequenceFlowTemplate;
-import io.camunda.operate.util.OperateZeebeAbstractIT;
-import io.camunda.operate.util.ZeebeTestUtil;
-import io.camunda.operate.webapp.rest.dto.SequenceFlowDto;
-import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
+import io.camunda.operate.entities.ProcessEntity;
+import io.camunda.operate.schema.indices.ProcessIndex;
+import io.camunda.operate.util.j5templates.OperateZeebeSearchAbstractIT;
+import java.io.IOException;
 import java.util.List;
-import org.junit.Test;
-import org.springframework.test.web.servlet.MvcResult;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 
-public class SequenceFlowZeebeIT extends OperateZeebeAbstractIT {
+public class MultiProcessZeebeImportIT extends OperateZeebeSearchAbstractIT {
 
-  protected String getSequenceFlowURL(Long processInstanceKey) {
-    return String.format(PROCESS_INSTANCE_URL + "/%s/sequence-flows", processInstanceKey);
-  }
+  @Autowired private ProcessIndex processIndex;
 
   @Test
-  public void testSequenceFlowsAreLoaded() throws Exception {
-    // having
-    final String processId = "demoProcess";
-    final BpmnModelInstance process =
-        Bpmn.createExecutableProcess(processId)
-            .startEvent("start")
-            .sequenceFlowId("sf1")
-            .serviceTask("task1")
-            .zeebeJobType("task1")
-            .sequenceFlowId("sf2")
-            .serviceTask("task2")
-            .zeebeJobType("task2")
-            .endEvent()
-            .done();
-    deployProcess(process, processId + ".bpmn");
+  public void shouldReadFromMultiProcessDiagram() throws IOException {
 
-    final Long processInstanceKey =
-        ZeebeTestUtil.startProcessInstance(
-            zeebeClient, processId, "{\"var1\": \"initialValue\", \"otherVar\": 123}");
-    searchTestRule.processAllRecordsAndWait(flowNodeIsActiveCheck, processInstanceKey, "task1");
-    ZeebeTestUtil.completeTask(zeebeClient, "task1", getWorkerName(), null);
-    searchTestRule.processAllRecordsAndWait(flowNodeIsActiveCheck, processInstanceKey, "task2");
+    // given
+    final String bpmnFile = "multi-process.bpmn";
+    operateTester.deployProcessAndWait(bpmnFile);
 
     // when
-    final List<SequenceFlowDto> sequenceFlows = getSequenceFlows(processInstanceKey);
+    final List<ProcessEntity> processEntities =
+        testSearchRepository.searchAll(processIndex.getAlias(), ProcessEntity.class);
+    final ProcessEntity process1 =
+        processEntities.stream()
+            .filter(x -> "process1".equals(x.getBpmnProcessId()))
+            .findAny()
+            .orElse(null);
+    final ProcessEntity process2 =
+        processEntities.stream()
+            .filter(x -> "process2".equals(x.getBpmnProcessId()))
+            .findAny()
+            .orElse(null);
 
-    assertThat(sequenceFlows)
-        .extracting(SequenceFlowTemplate.ACTIVITY_ID)
-        .containsExactlyInAnyOrder("sf1", "sf2");
-  }
+    // then
+    assertThat(processEntities).hasSize(2);
 
-  private List<SequenceFlowDto> getSequenceFlows(Long processInstanceKey) throws Exception {
-    final MvcResult mvcResult =
-        mockMvc
-            .perform(get(getSequenceFlowURL(processInstanceKey)))
-            .andExpect(status().isOk())
-            .andExpect(content().contentType(mockMvcTestRule.getContentType()))
-            .andReturn();
-    return mockMvcTestRule.listFromResponse(mvcResult, SequenceFlowDto.class);
+    assertThat(process1).isNotNull();
+    assertThat(process1.getName()).isEqualTo("Process 1");
+    assertThat(process1.getResourceName()).isEqualTo(bpmnFile);
+    assertThat(process1.getFlowNodes())
+        .extracting("id")
+        .containsExactlyInAnyOrder("startEvent1", "task1", "endEvent1");
+    assertThat(process1.getFlowNodes())
+        .extracting("name")
+        .containsExactlyInAnyOrder("Start Event 1", "Task 1", "End Event 1");
+
+    assertThat(process2).isNotNull();
+    assertThat(process2.getName()).isEqualTo("Process 2");
+    assertThat(process2.getResourceName()).isEqualTo(bpmnFile);
+    assertThat(process2.getFlowNodes())
+        .extracting("id")
+        .containsExactlyInAnyOrder("startEvent2", "task2", "endEvent2");
+    assertThat(process2.getFlowNodes())
+        .extracting("name")
+        .containsExactlyInAnyOrder("Start Event 2", "Task 2", "End Event 2");
   }
 }
