+++ operate/qa/integration-tests/src/test/java/io/camunda/operate/zeebeimport/SequenceFlowZeebeImportIT.java
@@ -16,78 +16,67 @@
  */
 package io.camunda.operate.zeebeimport;
 
-import static io.camunda.operate.qa.util.RestAPITestUtil.createGetAllProcessInstancesRequest;
+import static io.camunda.operate.webapp.rest.ProcessInstanceRestService.PROCESS_INSTANCE_URL;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.doReturn;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import io.camunda.operate.entities.FlowNodeInstanceEntity;
-import io.camunda.operate.property.OperateProperties;
+import io.camunda.operate.schema.templates.SequenceFlowTemplate;
 import io.camunda.operate.util.OperateZeebeAbstractIT;
-import io.camunda.operate.util.TestApplication;
 import io.camunda.operate.util.ZeebeTestUtil;
-import io.camunda.operate.webapp.reader.ListViewReader;
-import io.camunda.operate.webapp.rest.dto.listview.ListViewProcessInstanceDto;
-import io.camunda.operate.webapp.rest.dto.listview.ListViewRequestDto;
-import io.camunda.operate.webapp.rest.dto.listview.ListViewResponseDto;
-import io.camunda.operate.webapp.security.tenant.TenantService;
+import io.camunda.operate.webapp.rest.dto.SequenceFlowDto;
+import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
 import java.util.List;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MvcResult;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest(
-    classes = {TestApplication.class},
-    properties = {
-      OperateProperties.PREFIX + ".importer.startLoadingDataOnStartup = false",
-      OperateProperties.PREFIX + ".archiver.rolloverEnabled = false",
-      "spring.mvc.pathmatch.matching-strategy=ANT_PATH_MATCHER",
-      OperateProperties.PREFIX + ".multiTenancy.enabled = false"
-    })
-public class ImportMultitenancyDisabledZeebeIT extends OperateZeebeAbstractIT {
+public class SequenceFlowZeebeImportIT extends OperateZeebeAbstractIT {
 
-  @Autowired private ListViewReader listViewReader;
-
-  private String defaultTenantId = "<default>";
+  protected String getSequenceFlowURL(final Long processInstanceKey) {
+    return String.format(PROCESS_INSTANCE_URL + "/%s/sequence-flows", processInstanceKey);
+  }
 
   @Test
-  public void testDefaultTenantIsAssignedAndImported() {
-    doReturn(TenantService.AuthenticatedTenants.assignedTenants(List.of(defaultTenantId)))
-        .when(tenantService)
-        .getAuthenticatedTenants();
-
+  public void testSequenceFlowsAreLoaded() throws Exception {
     // having
     final String processId = "demoProcess";
-    final Long processDefinitionKey = deployProcess("demoProcess_v_1.bpmn");
+    final BpmnModelInstance process =
+        Bpmn.createExecutableProcess(processId)
+            .startEvent("start")
+            .sequenceFlowId("sf1")
+            .serviceTask("task1")
+            .zeebeJobType("task1")
+            .sequenceFlowId("sf2")
+            .serviceTask("task2")
+            .zeebeJobType("task2")
+            .endEvent()
+            .done();
+    deployProcess(process, processId + ".bpmn");
+
+    final Long processInstanceKey =
+        ZeebeTestUtil.startProcessInstance(
+            zeebeClient, processId, "{\"var1\": \"initialValue\", \"otherVar\": 123}");
+    searchTestRule.processAllRecordsAndWait(flowNodeIsActiveCheck, processInstanceKey, "task1");
+    ZeebeTestUtil.completeTask(zeebeClient, "task1", getWorkerName(), null);
+    searchTestRule.processAllRecordsAndWait(flowNodeIsActiveCheck, processInstanceKey, "task2");
 
     // when
-    final Long processInstanceKey =
-        ZeebeTestUtil.startProcessInstance(zeebeClient, processId, "{\"a\": \"b\"}");
-    searchTestRule.processAllRecordsAndWait(flowNodeIsActiveCheck, processInstanceKey, "taskA");
+    final List<SequenceFlowDto> sequenceFlows = getSequenceFlows(processInstanceKey);
 
-    // then
-    // assert process instance
-    final ListViewProcessInstanceDto pi = getSingleProcessInstanceForListView();
-    assertThat(pi.getTenantId()).isEqualTo(defaultTenantId);
-    // assert flow node instances
-    final List<FlowNodeInstanceEntity> allFlowNodeInstances =
-        tester.getAllFlowNodeInstances(processInstanceKey);
-    assertThat(allFlowNodeInstances).extracting("tenantId").containsOnly(defaultTenantId);
-    // assert variables
-
+    assertThat(sequenceFlows)
+        .extracting(SequenceFlowTemplate.ACTIVITY_ID)
+        .containsExactlyInAnyOrder("sf1", "sf2");
   }
 
-  private ListViewProcessInstanceDto getSingleProcessInstanceForListView(
-      ListViewRequestDto request) {
-    final ListViewResponseDto listViewResponse = listViewReader.queryProcessInstances(request);
-    assertThat(listViewResponse.getTotalCount()).isEqualTo(1);
-    assertThat(listViewResponse.getProcessInstances()).hasSize(1);
-    return listViewResponse.getProcessInstances().get(0);
-  }
-
-  private ListViewProcessInstanceDto getSingleProcessInstanceForListView() {
-    return getSingleProcessInstanceForListView(createGetAllProcessInstancesRequest());
+  private List<SequenceFlowDto> getSequenceFlows(final Long processInstanceKey) throws Exception {
+    final MvcResult mvcResult =
+        mockMvc
+            .perform(get(getSequenceFlowURL(processInstanceKey)))
+            .andExpect(status().isOk())
+            .andExpect(content().contentType(mockMvcTestRule.getContentType()))
+            .andReturn();
+    return mockMvcTestRule.listFromResponse(mvcResult, SequenceFlowDto.class);
   }
 }
